Floating point update rate, Arduino protocol 2.2
Use floating point for output device update rates, offering greater
precision at nearly no cost.

Move Arduino device to protocol 2.2, including parsing the decimal
update rate.  Protocol bump needed as int.parse() on floating point
throws a BadFormat exception.

+++ Actinic/OutputSystems/AbstractOutput.cs
@@ -55,7 +55,7 @@ namespace Actinic.Outputs
 		/// Length of time for an Update command to succeed (may be an average)
 		/// </summary>
 		/// <value>Delay in milliseconds</value>
-		public abstract int ProcessingLatency {
+		public abstract float ProcessingLatency {
 			get;
 		}
